Refactoring rule constant_012 into base rule multiline_array_alignment.
