+++ vsg/rules/__init__.py
@@ -67,6 +67,7 @@ from .existence_of_tokens_which_should_not_occur import existence_of_tokens_whic
 from .multiline_alignment_between_tokens import multiline_alignment_between_tokens
 from .multiline_structure import multiline_structure
 from .multiline_simple_structure import multiline_simple_structure
+from .multiline_array_alignment import multiline_array_alignment
 from .multiline_constraint_structure import multiline_constraint_structure
 from .multiline_procedure_call_structure import multiline_procedure_call_structure
 from .multiline_conditional_alignment import multiline_conditional_alignment
